Fixed -Wconversion and -Wshadow warnings. JIRA:CXX-5413
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@82500 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/ncbi_mbedtls.c
@@ -432,7 +432,7 @@ static int x_MbedTlsPush(void* ctx, const unsigned char* data, size_t size)
             } else {
                 assert(status == eIO_Success);
                 assert(x_written <= size);
-                n_written += x_written;
+                n_written += (ssize_t)x_written;
                 size      -= x_written;
                 data       = data + x_written;
             }
@@ -514,7 +514,7 @@ static EIO_Status s_MbedTlsRead(void* session, void* buf, size_t n_todo,
     } else
         status = eIO_Success;
 
-    *n_done = x_read;
+    *n_done = (size_t)x_read;
     return status;
 }
 
@@ -538,7 +538,7 @@ static EIO_Status x_MbedTlsWrite(void* session, const void* data,
     } else
         status = eIO_Success;
 
-    *n_done = x_written;
+    *n_done = (size_t)x_written;
     return status;
 }
